requirements.txt
pyrofork==2.3.69
tgcrypto==1.2.5
openai==1.58.1
aiohttp==3.11.11
loguru==0.7.3
pyyaml==6.0.2
aiosqlite==0.20.0
python-dotenv==1.0.1
